Dashboard: add A4A Marketplace > Exclusive offers screen, shared with a8c-for-agencies
Adds the Exclusive offers screen to the Multi-site Dashboard at /exclusive-offers,
nested under a new Marketplace sidebar group: an intro, DataViews search/filters
(Product / Offer type / Product type), and a responsive grid of partner offer
cards (Refer / Resell) with brand logos and CTAs.

The prop-driven PartnerOffers component + data + brand SVGs live in
client/dashboard/marketplace/exclusive-offers and are reused by
a8c-for-agencies, which injects its Redux tracking and the marketplace-type
sessionStorage CTA behavior. Removes the duplicated page components/constants/
types from a8c and repoints the learn section at the shared SVGs, so the brand
logos have a single source of truth.

Wiring: /exclusive-offers route, optional supports.marketplace flag, Marketplace
sidebar group, and the server path. Layout is fully component-driven (Grid /
Spacer); the only CSS is a8c's per-page DataViews stylesheet import.
