+++ client/a8c-for-agencies/sections/exclusive-offers/overview-content/style.scss
@@ -1,22 +1,8 @@
 @import "@wordpress/base-styles/mixins";
 
 @import '@wordpress/theme/design-tokens.css';
+
+// The shared dashboard PartnerOffers component renders DataViews. The dashboard
+// loads DataViews styles globally, but a8c-for-agencies loads them per-page, so
+// pull them in here for this page.
 @import '@wordpress/dataviews/build-style/style.css';
-
-.exclusive-offers-cards {
-	display: grid;
-	grid-template-columns: repeat( auto-fill, minmax( 250px, 1fr ) );
-	gap: 32px;
-
-	@include break-mobile {
-		grid-template-columns: repeat( auto-fill, minmax( 270px, 1fr ) );
-	}
-
-	@include break-huge {
-		grid-template-columns: repeat( auto-fill, minmax( 300px, 1fr ) );
-	}
-}
-
-.exclusive-offers-filters {
-	margin-bottom: 16px;
-}
