+++ client/a8c-for-agencies/sections/learn/format-agency-resources.tsx
@@ -1,8 +1,8 @@
-import JetpackLogo from 'calypso/a8c-for-agencies/sections/exclusive-offers/overview-content/images/jetpack.svg';
-import PressableLogo from 'calypso/a8c-for-agencies/sections/exclusive-offers/overview-content/images/pressable.svg';
-import VIPLogo from 'calypso/a8c-for-agencies/sections/exclusive-offers/overview-content/images/vip.svg';
-import WooLogo from 'calypso/a8c-for-agencies/sections/exclusive-offers/overview-content/images/woo.svg';
-import WordPressDotComLogo from 'calypso/a8c-for-agencies/sections/exclusive-offers/overview-content/images/wordpressdotcom.svg';
+import JetpackLogo from 'calypso/dashboard/marketplace/exclusive-offers/images/jetpack.svg';
+import PressableLogo from 'calypso/dashboard/marketplace/exclusive-offers/images/pressable.svg';
+import VIPLogo from 'calypso/dashboard/marketplace/exclusive-offers/images/vip.svg';
+import WooLogo from 'calypso/dashboard/marketplace/exclusive-offers/images/woo.svg';
+import WordPressDotComLogo from 'calypso/dashboard/marketplace/exclusive-offers/images/wordpressdotcom.svg';
 import type { APIAgencyResource } from 'calypso/a8c-for-agencies/data/learn/types';
 import type { ResourceItem } from 'calypso/a8c-for-agencies/sections/learn/resource-center/overview-content/types';
 import type { JSX } from 'react';
